# docs/openapi.yaml
openapi: 3.0.3
info:
  title: "ASE Project API"
  version: "1.0.0"
  description: "API documentation for ASE Project"
servers:
  - url: "http://localhost:5050"
paths:
...
  /api/player/gacha/player-collection:
    get:
      summary: "Get player's gacha collection"
      tags: ["GachaMicroservice"]
      responses:
        '200':
          description: "Player's gacha collection"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GachaCollectionItem'
        '404':
          description: "User not found"
...
  /api/player/gacha/player-collection/gacha/{gacha_id}:
    get:
      summary: "Get specific gacha item details for a player"
      tags: ["GachaMicroservice"]
      parameters:
        - in: path
          name: gacha_id
          required: true
          schema:
            type: integer
          description: "ID of the gacha item"
      responses:
        '200':
          description: "Details of the specific gacha item for a player"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GachaItem'
        '404':
          description: "Gacha item not found"
...
  /api/player/gacha/roll:
    post:
      summary: "Roll a new gacha item"
      tags: ["GachaMicroservice"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  description: "ID of the user rolling the gacha"
      responses:
        '200':
          description: "Roll successful"
          content:
            application/json:
              schema:
                type: object
                properties:
                  collectionId:
                    type: integer
                    description: "ID of the added item in user's collection"
        '404':
          description: "User not found"
        '400':
          description: "Rolling failed, for example the user's in-game balance is not sufficient"
...
  /api/player/currency:
    post:
      summary: "Purchase in-game currency"
      tags: ["PaymentsMicroservice"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurrencyPurchase'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Purchase successful"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  transaction:
                    $ref: '#/components/schemas/Transaction'
        '400':
          description: "Invalid request data"
        '404':
          description: "User not found"
        '500':
          description: "Internal server error"
  /api/player/decrease/{user_id}:
    put:
      summary: "Decrease player's in-game currency"
      tags: ["PaymentsMicroservice"]
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: "ID of the user"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurrencyChange'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Currency decreased successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  transaction:
                    $ref: '#/components/schemas/Transaction'
        '400':
          description: "Invalid decrease request or insufficient balance"
        '404':
          description: "User not found"
        '500':
          description: "Internal server error"
...
  /api/player/register:
    post:
      summary: "Register a new player"
      tags: ["Player"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPlayerRequest'
      responses:
        '200':
          description: "Player registered successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerResponse'
        '400':
          description: "Invalid input or registration failed"
        '409':
          description: "Username already taken"
  /api/player/login:
    post:
      summary: "Login a player"
      tags: ["Player"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: "Login successful"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: "Invalid username or password"
        '401':
          description: "Unauthorized access"
  /api/player/logout:
    post:
      summary: "Log out a player"
      tags: ["Player"]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Player successfully logged out"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: "Error during logout"
        '401':
          description: "Unauthorized access"
...
  /player/auction/market:
    get:
      summary: Get active auctions
      description: Retrieve all active auctions.
      tags: [AuctionMicroservice]
      responses:
        "200":
          description: List of active auctions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuctionItem"
        "500":
          description: Internal Server Error
  /player/auction/create:
    post:
      summary: Create a new auction
      description: Create a new auction listing for a gacha item.
      tags:
        - Player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gachaCollectionId:
                  type: integer
                auctionStart:
                  type: string
                  format: date-time
                auctionEnd:
                  type: string
                  format: date-time
                minimumBid:
                  type: number
              required:
                - gachaCollectionId
                - auctionStart
                - auctionEnd
                - minimumBid
      responses:
        "201":
          description: Auction created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  auctionId:
                    type: integer
        "400":
          description: Invalid input data
        "500":
          description: Internal Server Error
  /player/auction/bid/{auction_id}:
    post:
      summary: Place a bid on an auction
      description: Place a bid on an active auction.
      tags:
        - Player
      parameters:
        - name: auction_id
          in: path
          required: true
          description: ID of the auction
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bidAmount:
                  type: number
              required:
                - bidAmount
      responses:
        "200":
          description: Bid placed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  bidId:
                    type: integer
        "400":
          description: Invalid request
        "500":
          description: Internal Server Error
...
components:  
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GachaItem:
      type: object
      properties:
        id:
          type: integer
          description: "ID of the gacha item"
        name:
          type: string
          description: "Name of the gacha item"
        description:
          type: string
          description: "Description of the gacha item"
        image:
          type: string
          description: "URL to the gacha item's image"
        rarityPercent:
          type: number
          format: float
          description: "Rarity percentage of the gacha item"
...
    GachaCollectionItem:
      type: object
      properties:
        id:
          type: string
          description: "Name of the gacha item"
        gachaId:
          type: string
          description: "Description of the gacha item"
        userId:
          type: string
          description: "URL to the gacha item's image"
        source:
          type: number
          format: float
          description: "Rarity percentage of the gacha item"
        timestamp:
          type: string
          description: "Creation timestamp"
    CurrencyPurchase:
      type: object
      properties:
        amount:
          type: number
          format: float
          description: "Amount of currency to purchase"
      required:
        - amount
    CurrencyChange:
      type: object
      properties:
        amount:
          type: number
          format: float
          description: "Amount to add or subtract"
      required:
        - amount
    Transaction:
      type: object
      properties:
        id:
          type: integer
          description: "Transaction ID"
        userId:
          type: integer
          description: "User ID for the transaction"
        amount:
          type: number
          format: float
          description: "Transaction amount"
        timestamp:
          type: string
          format: date-time
          description: "Transaction timestamp"
...
    RegisterPlayerRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        profilePicture:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
    PlayerResponse:
      type: object
      properties:
        userId:
          type: integer
...
    AuctionItem:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the auction.
        gachaCollectionId:
          type: integer
          description: Identifier for the associated Gacha collection.
        userId:
          type: integer
          description: Identifier for the user associated with the auction.
        auctionStart:
          type: string
          format: date-time
          description: Start date and time of the auction.
        auctionEnd:
          type: string
          format: date-time
          description: End date and time of the auction.
        minimumBid:
          type: number
          format: float
          description: Minimum bid amount required for the auction.
        timestamp:
          type: string
          format: date-time
          description: Timestamp for when the auction was created or updated.
        status:
          type: string
          description: Current status of the auction.
          enum:
            - ACTIVE
            - PASSED
            - UNKNOWN
...
security:
  - bearerAuth: []